+++ test/optimization/test_mode_estimators.jl
@@ -51,10 +51,10 @@ using Optim, OptimizationOptimJL
     end
 
 
-    @testset "ModeAsDefined" begin
+    @testset "PredefinedMode" begin
         context = BATContext()
-        @test @inferred(bat_findmode(prior, ModeAsDefined(), context)).result == true_mode
-        @test @inferred(bat_findmode(batmeasure(prior), ModeAsDefined(), context)).result == true_mode
+        @test @inferred(bat_findmode(prior, PredefinedMode(), context)).result == true_mode
+        @test @inferred(bat_findmode(batmeasure(prior), PredefinedMode(), context)).result == true_mode
         let post_modes = @inferred(bat_findmode(posterior, context)).result
             for k in keys(post_modes)
                 @test isapprox(post_modes[k], true_mode[k], atol=0.001)
@@ -63,10 +63,10 @@ using Optim, OptimizationOptimJL
     end
 
 
-    @testset "MaxDensitySearch" begin
+    @testset "EmpiricalMode" begin
         context = BATContext()
-        @test @inferred(bat_findmode(samples, MaxDensitySearch(), context)).result isa NamedTuple
-        m = bat_findmode(samples, MaxDensitySearch(), context)
+        @test @inferred(bat_findmode(samples, EmpiricalMode(), context)).result isa NamedTuple
+        m = bat_findmode(samples, EmpiricalMode(), context)
         @test samples[m.mode_idx].v == m.result
         @test isapprox(unshaped(m.result, elshape(samples.v)), true_mode_flat, rtol = 0.05)
     end